BugFix Update
Player/Ban Manager
These two files write to or check against the database.
Removed redundant .toString(); call.

Removed unused variables.

KickManager recieves a bit of an overhaul, relying less on the main
code, for readability.
Because of the data of the boolean useM, I can shorten to a single
broadcastMessage() which varies directly within it's single line.